use linearise
